bind offset-sub to Alt+o
Because it is consistent with the other Alt bindings, and I have a
script to enter a path to load with Ctrl+o which is consistent with the
standard GUI keybinding.

Incidentally, musixmatch-download isn't bound to Ctrl+m because that
doesn't work in the terminal, and not to Shift+m because I already have
that bound to show-text ${metadata} and because Shift+o is already bound
by default.